Better round trip for RequestContext<->Dict conversion
Conversion of nova.context.RequestContext to and from a
dictionary was not working properly. There is duplicate
information as well in the serialized dict. As a first
step i've removed request_id as it dups the one in
oslo_context.RequestContext. However changing project_id
and user_id to a @property is challenging at this late
juncture. So leaving them as-is with a FIXME. The other
issue related to the roundtrip is the stack traces that
show up in screen-n-cpu log as mentioned in the bug
referenced below. To avoid that added some defensive
code ('has_attr') for now and a FIXME.

Partial-Bug: #1427665
Change-Id: Ia47d4909d2656d6fc4c1179659b8098bba3235d3